+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # JXWord Creator Changelog
 
+## 0.2.4
+
+- Allow customisation of download filename
+- Set filename to uploaded file name
+- Only show patterns if we have some for the size
+- Some formatting
+
 ## 0.2.3
 
 - Start Changelog